fix: based on the review
