+++ components/checkout/plan/SelectablePlan.vue
@@ -7,7 +7,6 @@ import { formatCurrency } from '~/utils/text';
 const props = withDefaults(
   defineProps<{
     plan: AvailablePlan;
-    isMostPopular?: boolean;
     selected?: boolean;
     period: PricingPeriod;
     readonly?: boolean;
@@ -15,7 +14,6 @@ const props = withDefaults(
   }>(),
   {
     selected: false,
-    isMostPopular: false,
   },
 );
 
@@ -26,7 +24,7 @@ const emit = defineEmits<{
 
 const { t } = useI18n({ useScope: 'global' });
 
-const { plan, period, selected, isMostPopular } = toRefs(props);
+const { plan, period, selected } = toRefs(props);
 
 const price = computed<string | undefined>(() => {
   const { monthlyPlan, yearlyPlan } = get(plan);
@@ -56,7 +54,7 @@ watch(price, (price) => {
     @click="emit('click')"
   >
     <RuiChip
-      v-if="isMostPopular"
+      v-if="plan.isMostPopular"
       color="primary"
       class="mb-2 !min-h-4"
       :class="{
